LayoutTests/accessibility/mac/popup-button-exposes-axvalue-expected.txt

This tests that pop up buttons expose the AXValue attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS popup.isAttributeSupported('AXValue') is true
PASS ariaPopup.isAttributeSupported('AXValue') is true
PASS successfullyParsed is true

TEST COMPLETE
